.github/skills/sample_app_components/mcp_server/pyproject.toml
[project]
name = "af_mcp_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp",
    "azure-search-documents",
    "azure-identity",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx[socks,http2]>=0.27,<0.29",
    "httpx-sse>=0.4.0,<0.5",
    "psycopg",
    "psycopg_pool",
    "psycopg_binary"
]

# Safer to omit prereleases globally. Remove this block, or set to disallow.
[tool.uv]
prerelease = "allow"
